[ASSIGNMENT 2]
- Thanks to Erik from the 3dgep for providing parts of the answer/code, couldn't figure everything out by myself.
- Overloaded Surface::Plot to take floating-point variables.
- Added Surface::GetSubPixelValue to get the, well, subpixel value.
- Made subpixel rendering possible with the current code.

+++ game.cpp
@@ -26,7 +26,7 @@ namespace Tmpl8
             float dist = sqrtf(dx * dx + dy * dy);
             if (dist < 50)
                 x[i] += dx / dist, y[i] += dy / dist;
-            screen->Plot((int)x[i], (int)y[i], 0xffffff);
+            screen->Plot(x[i], y[i], 0xffffff);
         }
     }
 };
